# backend-rust/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ABP Studio Rust API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/simple-batch-generate": {
      "post": {
        "operationId": "simple_batch_generate",
        "summary": "Migrated from Python simple_batch_generate",
        "responses": {
          "200": {
            "description": "Compatible JSON response"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
  }
}